Rename references to new directory
