+++ agents/exec/README.md
@@ -23,7 +23,7 @@ Development
 ```bash
 export CHE_PATH=~/code/che
 mkdir $GOPATH/src/github.com/eclipse/che/agents -p
-ln -s $CHE_PATH/agents/exec-agent/src $GOPATH/src/github.com/eclipse/che/agents/exec-agent
+ln -s $CHE_PATH/agents/exec/src $GOPATH/src/github.com/eclipse/che/agents/exec
 ```
 
 ##### Install godep
@@ -34,26 +34,26 @@ go get github.com/tools/godep
 ##### Get all dependencies
 
 ```bash
-cd $GOPATH/src/github.com/eclipse/che/agents/exec-agent
+cd $GOPATH/src/github.com/eclipse/che/agents/exec
 $GOPATH/bin/godep restore
 ```
 
-That's it, `$GOPATH/src/github.com/eclipse/che/agents/exec-agent` project is ready.
+That's it, `$GOPATH/src/github.com/eclipse/che/agents/exec` project is ready.
 
 ##### Building linked project
 
 ```bash
-cd $GOPATH/src/github.com/eclipse/che/agents/exec-agent && go build
+cd $GOPATH/src/github.com/eclipse/che/agents/exec && go build
 ```
 
 ##### Running linked project tests
 
 ```bash
-cd $GOPATH/src/github.com/eclipse/che/agents/exec-agent && go test ./...
+cd $GOPATH/src/github.com/eclipse/che/agents/exec && go test ./...
 ```
 
 ##### Formatting linked project sources
 
 ```bash
-cd $GOPATH/src/github.com/eclipse/che/agents/exec-agent && go fmt ./...
+cd $GOPATH/src/github.com/eclipse/che/agents/exec && go fmt ./...
 ```
