+++ agents/exec/src/Godeps/Godeps.json
@@ -1,5 +1,5 @@
 {
-	"ImportPath": "github.com/eclipse/che/agents/exec-agent",
+	"ImportPath": "github.com/eclipse/che/agents/exec",
 	"GoVersion": "go1.6",
 	"GodepVersion": "v74",
 	"Deps": [
