+++ agents/exec/src/process/logs_reader_test.go
@@ -12,7 +12,7 @@
 package process_test
 
 import (
-	"github.com/eclipse/che/agents/exec-agent/process"
+	"github.com/eclipse/che/agents/exec/process"
 	"os"
 	"testing"
 	"time"
