+++ agents/exec/src/process/process.go
@@ -21,7 +21,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/eclipse/che/agents/exec-agent/rpc"
+	"github.com/eclipse/che/agents/exec/rpc"
 )
 
 const (
