+++ agents/exec/src/process/process_cleaner_test.go
@@ -14,7 +14,7 @@ package process_test
 import (
 	"testing"
 
-	"github.com/eclipse/che/agents/exec-agent/process"
+	"github.com/eclipse/che/agents/exec/process"
 	"time"
 )
